Initial commits to replace atomic
